move listener/listener.go to decoder/listener.go
