+++ decoder/listener.go
@@ -2,18 +2,18 @@
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-package listener
+package decoder
 
 import "github.com/muktihari/fit/proto"
 
-// MesgListener is an interface to listen to message decoded event.
+// MesgListener is an interface for listening to message decoded events.
 type MesgListener interface {
-	// OnMesg receives message from emitter.
+	// OnMesg receives message from Decoder.
 	OnMesg(mesg proto.Message)
 }
 
-// MesgDefListener is an interface to listen to message definition decoded event.
+// MesgDefListener is an interface for listening to message definition decoded event.
 type MesgDefListener interface {
-	// OnMesgDef receives message definition from emitter.
+	// OnMesgDef receives message definition from Decoder.
 	OnMesgDef(mesgDef proto.MessageDefinition)
 }
