--- listener/doc.go
@@ -1,7 +1,0 @@
-// Copyright 2023 The FIT SDK for Go Authors. All rights reserved.
-// Use of this source code is governed by a BSD-style
-// license that can be found in the LICENSE file.
-
-// Package listener defines interface contracts that can be added on a decoder to receive message and message definition
-// as soon as it is decoded.
-package listener
